send feedback
Added two sites, reporting_textarea.php, which allows the user to write feedback into a text box. This feedback is written into a file on the server and is unaccessible for users, because of Debian webserver settings, which disallow users to access files without an extension. As you probably have guessed, the feedback is saved into a file without an extension. This log contains:
-time and date,
-feedback,
-IP address of sender.

KNOW VULNERABILITES:
This is unsecure and allows for remote code execution because of the textarea box. I currently have no idea on how to fix this, as I have little experience with webservers.